Preload email templates during initialization
Load templates in parallel with connection check and custom parameters
during component initialization. The loading screen now waits for both
mode determination and templates to be ready before rendering the main UI.

Updated TemplatesModal to receive templates as props instead of loading
them when opened, eliminating the delay when opening the templates modal.
